Try to implement a generic way for doWriteAfter and doWriteBefore
